change dashboard

+++ src/main/resources/static/app/project/activity-progress/progress.js
@@ -3,24 +3,29 @@
 
     angular.module('app.project')
         .controller('Progress', Progress);
-    Progress.$inject = ['reportservice','progressservice', 'trainingcenterservice', '$scope', 'NgTableParams', 'logger', '$routeParams'];
+    Progress.$inject = ['activityservice', 'reportservice', 'progressservice', 'trainingcenterservice', '$scope', 'NgTableParams', 'logger', '$routeParams'];
 
-    function Progress(reportservice,progressservice, trainingcenterservice, $scope, NgTableParams, logger, $routeParams) {
+    function Progress(activityservice, reportservice, progressservice, trainingcenterservice, $scope, NgTableParams, logger, $routeParams) {
         var vm = this;
         var self = this;
         var activityId = $routeParams.activityId;
         // alert(pCode);
-        if(activityId!=undefined){
-            vm.forReport=true;
+        if (activityId != undefined) {
+            vm.forReport = true;
+            activityservice.getSingleActivity({id: activityId}).$promise.then(function (data) {
+                vm.title = "Progress View Panel of ' Activity ' : " + "' " + data.activityName + " '";
 
-        $scope.timeFrame = ["FIRST_QUARTER", "SECOND_QUARTER", "THIRD_QUARTER", "SELECT"];
-        $scope.selectedTimeFrameItem = $scope.timeFrame[3];
-        $scope.timeFrame.pop();
-        findAllTC();
-        findAllActivitiesProgressByActivityId(activityId);
+            });
+
+            $scope.timeFrame = ["FIRST_QUARTER", "SECOND_QUARTER", "THIRD_QUARTER", "SELECT"];
+            $scope.selectedTimeFrameItem = $scope.timeFrame[3];
+            $scope.timeFrame.pop();
+            findAllTC();
+            findAllActivitiesProgressByActivityId(activityId);
         }
-        else{
-            vm.forReport=false;
+        else {
+            vm.title = "Progress View Panel";
+            vm.forReport = false;
             $scope.fYear = ['2072-73', '2073-74', '2074-75', '2075-76', '2077-78'];
             $scope.selectedItem = $scope.fYear[0];
             findAllProgressesByFiscalYear({fiscalYear: $scope.selectedItem});
@@ -33,13 +38,23 @@
         }
 
         $scope.initCreatePanel = function () {
-            // $scope.clearValidationMessages();
+            $scope.clearValidationMessages();
             vm.riskyId = 0;
 
 
             vm.title = "Create Progress Panel";
             vm.showCreatePanel = !vm.showCreatePanel;
             $scope.btnText = "Create ";
+        }
+        $scope.clearValidationMessages = function () {
+            $scope.progressmodel = {};
+            $scope.progressmodel.description = "";
+            $scope.progressmodel.progressQty = "";
+            $scope.progressmodel.goalQty = "";
+            $scope.timeFrame.push("SELECT");
+            $scope.selectedTimeFrameItem = $scope.timeFrame[3];
+
+
         }
         $scope.closeThePanel = function () {
             vm.showCreatePanel = !vm.showCreatePanel;
@@ -76,9 +91,12 @@
 
         }
         $scope.deleteThisProgress = function (pid) {
-            progressservice.deleteProgress({id: pid}).$promise.then(function (data) {
-                findAllActivitiesProgressByActivityId(activityId);
-            })
+            if (confirm("Are you sure you want to delete this progress ?")) {
+                progressservice.deleteProgress({id: pid}).$promise.then(function (data) {
+                    findAllActivitiesProgressByActivityId(activityId);
+                })
+
+            }
         }
         function showValidationErrors() {
             $scope.submitted = true;
@@ -153,8 +171,8 @@
 
         }
 
-        function findAllProgressesByFiscalYear(fYear){
-            reportservice.findAllProgresses(fYear).$promise.then(function(data){
+        function findAllProgressesByFiscalYear(fYear) {
+            reportservice.findAllProgresses(fYear).$promise.then(function (data) {
                 self.tableParams = new NgTableParams({}, {dataset: data});
 
             })
