+++ src/main/resources/static/app/training/training.js
@@ -3,9 +3,9 @@
 
     angular.module('app.training')
         .controller('Training', Training);
-    Training.$inject = ['trainingservice', '$scope', 'NgTableParams', 'logger', '$routeParams'];
+    Training.$inject = ['trainingcenterservice','trainingservice', 'memberservice', '$scope', 'NgTableParams', 'logger', '$routeParams'];
 
-    function Training(trainingservice, $scope, NgTableParams, logger, $routeParams) {
+    function Training(trainingcenterservice,trainingservice, memberservice, $scope, NgTableParams, logger, $routeParams) {
         var tcId = $routeParams.id;
         findAll(tcId);
 
@@ -13,15 +13,45 @@
 
         var vm = this;
 
-        vm.riskyId=0;
+        vm.riskyId = 0;
 
         $scope.tmodel = {};
+        trainingcenterservice.getTrainingCenter({id:tcId}).$promise.then(function(data){
+            vm.mainTitle = "Trainings View Panel of ' Training Center ' : " + "' "+data.name+" '" ;
+
+        });
+
         $scope.initCreatePanel = function () {
-            vm.riskyId=0;
+            $scope.clearValidationMessages();
+            vm.riskyId = 0;
             vm.title = "Create Training Panel ";
             vm.showCreatePanel = !vm.showCreatePanel;
             $scope.btnText = "Create ";
         }
+
+
+        $scope.clearValidationMessages = function () {
+            $scope.submitted = false;
+
+            //clear validation messages for username
+            $scope.tmodel.name = "";
+            $scope.tmodel.budget ="";
+            $scope.tmodel.start ="";
+            $scope.tmodel.end ="";
+            $scope.tmodel.target = "";
+            $scope.tmodel.duration = "";
+            $scope.tForm.duration.$pristine = true;
+            $scope.tForm.name.$pristine = true;
+            $scope.tForm.budget.$pristine = true;
+            $scope.tForm.start.$pristine = true;
+            $scope.tForm.end.$pristine = true;
+            $scope.tForm.target.$pristine = true;
+        }
+
+
+
+
+
         $scope.initEditPanel = function (pid) {
             $scope.btnText = "Update";
             vm.riskyId = pid;
@@ -55,8 +85,7 @@
                 start: $scope.tmodel.start,
                 end: $scope.tmodel.end,
                 target: $scope.tmodel.target,
-                trainingCenterId:tcId,
-
+                trainingCenterId: tcId,
 
 
             }
@@ -65,9 +94,21 @@
 
         }
         $scope.deleteTheTraining = function (trainingId) {
-            trainingservice.deleteTraining({id: trainingId}).$promise.then(function (data) {
-                findAll(tcId);
+            memberservice.findAllMembers({id:trainingId}).$promise.then(function (data) {
+                if (data.length > 0) {
+                    alert("Cannot delete ! Has memberes in it")
+                }
+                else {
+                    if (confirm("Are you sure you want to delete ?")) {
+
+                        trainingservice.deleteTraining({id: trainingId}).$promise.then(function (data) {
+                            findAll(tcId);
+                        });
+                    }
+
+                }
             });
+
         }
         function createOrEditTraining(tc) {
             trainingservice.addTraining(tc).$promise.then(function (data) {
